task 6 lesson 01 amended on teacher remark

+++ Homework/Lesson01/task6/Program.cs
@@ -3,7 +3,7 @@
 Console.WriteLine("Enter integral number: ");
 int number = Convert.ToInt32(Console.ReadLine());
 
-double result = number % 2;
+int result = number % 2;
 
 if (result == 0) 
 {
